owner|refactor category screen
