--- owner/src/main/java/org/the_chance/honeymart/ui/addCategory/AddCategoryUIEffect.kt
@@ -1,4 +1,0 @@
-package org.the_chance.honeymart.ui.addCategory
-
-class AddCategoryUIEffect() {
-}
