+++ owner/src/main/java/org/the_chance/honeymart/ui/features/category/CategoriesScreen.kt
@@ -19,12 +19,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
-import org.the_chance.honeymart.ui.features.category.categories.AddCategoryContent
-import org.the_chance.honeymart.ui.addCategory.composable.EmptyCategory
-import org.the_chance.honeymart.ui.addCategory.composable.HoneyMartTitle
-import org.the_chance.honeymart.ui.features.category.categories.CategoryItems
-import org.the_chance.honeymart.ui.features.category.categories.CategoryProducts
-import org.the_chance.honeymart.ui.features.category.categories.UpdateCategoryContent
+import org.the_chance.honeymart.ui.features.category.composable.AddCategoryContent
+import org.the_chance.honeymart.ui.features.category.composable.EmptyCategory
+import org.the_chance.honeymart.ui.features.category.composable.HoneyMartTitle
+import org.the_chance.honeymart.ui.features.category.composable.CategoryItems
+import org.the_chance.honeymart.ui.features.category.composable.CategoryProducts
+import org.the_chance.honeymart.ui.features.category.composable.UpdateCategoryContent
 import org.the_chance.honymart.ui.composables.ConnectionErrorPlaceholder
 import org.the_chance.honymart.ui.composables.CustomAlertDialog
 import org.the_chance.honymart.ui.composables.Loading
